requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
jamo
pydantic
hgtk
six
pyahocorasick